First update of the Getting started instructions
Pranali - 

I used some of the existing getting started content to fill in the gaps. Let me know if it is valid or if v4 is a different process. I have inline questions in the .md text. You can reply by email or in the comments on the Pul request.

Thanks, 

Jim